Minor issue fixes and test reorganisation

LGTM

+++ test/test_dtypes_integration.py
@@ -7,17 +7,17 @@ odML parsers.
 import datetime as dt
 import os
 import shutil
-import tempfile
 import unittest
 
 import odml
+from .util import create_test_dir
 
 
 class TestTypesIntegration(unittest.TestCase):
 
     def setUp(self):
         # Set up test environment
-        self.tmp_dir = tempfile.mkdtemp(suffix=".odml")
+        self.tmp_dir = create_test_dir(__file__)
 
         self.json_file = os.path.join(self.tmp_dir, "test.json")
         self.xml_file = os.path.join(self.tmp_dir, "test.xml")
@@ -29,7 +29,7 @@ class TestTypesIntegration(unittest.TestCase):
         self.doc = doc
 
     def tearDown(self):
-        if os.path.exists(self.tmp_dir):
+        if self.tmp_dir and os.path.exists(self.tmp_dir):
             shutil.rmtree(self.tmp_dir)
 
     def test_time(self):
